debian/control: mark some packages Multi-Arch: same
The packages kopano-contacts, kopano-dev and kopano-libs are
'Architecture: any' and don't have any postinst or postrm scripts, this
means these packages are equal for all platforms so we can mark them
'Multi-Arch: same'.
This needs to be removed once the packages are getting specific for
various platforms or get additional postinst or -rm scripts.

https://wiki.debian.org/MultiArch/Hints#ma-same