debian/control: let kopano-dev conflicting with libmapi-dev
The package libmapi-dev from src:openchange is also providing a library
libmapi.so. Unfortunately src:kopanocore is also providing currently a
library with the same name.
Jelmer Vernooij as one of the OpenChange maintainers confirmed that
upstream development of OpenChange is mostly dead and the binary
packages are kicked out of testing and unstable. Adding a Conflicts
against libmapi-dev should prevent errors before dpkg is giving up and
users need to interact manually.

Future versions of kopanocore will come with a different naming for the
Kopano specific libraries and this marking could be removed then.

Closes: #857377