lintian: adding additional override for kopano-libs
kopano-libs is a package that is providing all the various libraries
that are built from source and needed at least from the various parts of
kopanocore binaries.
But this is lintian complaining about, the package name doesn't fit to a
library name but we can ignore this hint. It would be a bit overkill to
package every library into a own binary package.