d/rules: detect PHP_VERSION dynamically
In commit 5df0116 we switch over to use php-dev as the package we depned
on while building instead of php$VERSION-dev. And in commit 309dc85 we
have taken the API date of the PHP version dynamically from the
environment.
This commit is taking the same approach for the PHP version itself so we
don't need (hopefully) ever to modify any PHP releated information
within the rules file.