autopkgtests: make the test working again
Remove all the tweeking things and rework the smoke test to get it
working the "normal" way.
Due some changes in the past like adding and reordering of tests the old
autopkgtest setup wasn't working anymore as usual. Mainly because
dbconfig-common was running while no mariadb-server was interacting.

Reorder the tests which get called in the control file so we first check
the functionality of the kopano-{utils,server} packages (again).