d/kopano-common: remove /v/log/kopano on purging
The postinst of kopano-common is modifying the folder /var/log/kopano
while the configure run for kopano-common. The folder itself is getting
installed by the package itself.
If user is selecting the purging option while removing kopano-common we
can remove the complete folder with all subfolders to. This can be done
by the postrm script.