d/kopano-utils.install: change substitution variable
Obviously some kind of an oversight in the past while working on
versions after 8.6.5 ...
There is the substitution variable '@DEB_HOST_MULTIARCH@' inside the install
sequencer file for kopano-utils used and we can't use the dedicated special
target within debian/rules for finding all *.in files and replace the
variable '@DEB_HOST_MULTIARCH@' inside these files with the correct
value without further modification of this target (dh-exec needs
execution rights on the file and we would need to ensure the correct
rights on the new file after processing) we don't use a *.in file here.
But we need to change '@DEB_HOST_MULTIARCH@' to '${DEB_HOST_MULTIARCH}'
then.

Closes: #920189