apparmor kopano-search: tune existing profile more
Adding new rule for specific binaries in the folder {/usr}/bin which
doesn't need to have full access on the system.

Adding read access right to /etc/mime.types.

And finally allow the inclusion of a local configurated AA profile by
the system user so he can overrule something without the modification of
a file within the packaging.

Closes: #934460