autopkgtest: adjust and expand the automatic testing
By the recently added various upstream commits as additional patches for
the patch-queue the kopano-server is acting a bit different than
before.

If a user is newly added to the server that user will also get a store
automatically. Thus we don't need to create a store for that user within
the autopkgtest. Instead we simply can check if the server can provide
details for the user.

But we can also check if we can create a public store on the system.